requirements.txt
absl-py==0.7.1
astor==0.8.0
Click==7.0
Flask==1.1.0
Flask-Sijax==0.4.1
future==0.17.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
nltk==3.4.4
numpy==1.16.4
pandas==0.24.2
protobuf==3.8.0
python-dateutil==2.8.0
pytz==2019.1
Sijax==0.3.2
six==1.12.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
Werkzeug==0.15.4
wrapt==1.11.2
